fix: updated some API changes for the future RL release.

+++ src/main/java/com/twitchliveloadout/marketplace/spawns/SpawnUtilities.java
@@ -57,16 +57,16 @@ public class SpawnUtilities {
 
 		double angleRadians = Math.toRadians(angleDegrees);
 		for(int verticesIndex = 0; verticesIndex < modelData.getVerticesCount(); ++verticesIndex) {
-			int[] xVertices = modelData.getVerticesX();
-			int[] yVertices = modelData.getVerticesY();
-			int[] zVertices = modelData.getVerticesZ();
-			int x = xVertices[verticesIndex];
-			int y = yVertices[verticesIndex];
-			int z = zVertices[verticesIndex];
+			float[] xVertices = modelData.getVerticesX();
+			float[] yVertices = modelData.getVerticesY();
+			float[] zVertices = modelData.getVerticesZ();
+			float x = xVertices[verticesIndex];
+			float y = yVertices[verticesIndex];
+			float z = zVertices[verticesIndex];
 
-			xVertices[verticesIndex] = (int) (x*Math.cos(angleRadians) + z*Math.sin(angleRadians));
+			xVertices[verticesIndex] = (float) (x*Math.cos(angleRadians) + z*Math.sin(angleRadians));
 			yVertices[verticesIndex] = y;
-			zVertices[verticesIndex] = (int) (z*Math.cos(angleRadians) - x*Math.sin(angleRadians));
+			zVertices[verticesIndex] = (float) (z*Math.cos(angleRadians) - x*Math.sin(angleRadians));
 		}
 	}
 }
